# it-pagalbos-biletai/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IT Pagalbos Biletai API",
        "description": "REST API for workplaces, devices, and tickets",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/v1/devices": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "List all devices (with optional filters)",
                "operationId": "3c61e6c1cd697bc8a6d85415c7d9567c",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pc",
                                "printer",
                                "network",
                                "other"
                            ]
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive",
                                "maintenance"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful list"
                    }
                }
            },
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Create a new device",
                "operationId": "19839aad1adbee27ad02ff85d89b0764",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "workplace_id",
                                    "type",
                                    "status"
                                ],
                                "properties": {
                                    "workplace_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "pc",
                                            "printer",
                                            "network",
                                            "other"
                                        ],
                                        "example": "pc"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "maintenance"
                                        ],
                                        "example": "active"
                                    },
                                    "serial": {
                                        "type": "string",
                                        "example": "SN-1234-XYZ"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "400": {
                        "description": "Bad payload"
                    }
                }
            }
        },
        "/api/v1/devices/{id}": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Get a device by ID",
                "operationId": "d86fe9541c8fba92445b60b004be2036",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Found"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Devices"
                ],
                "summary": "Update a device",
                "operationId": "51c74e084d6f2a9253840a2f561e26c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "pc",
                                            "printer",
                                            "network",
                                            "other"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "maintenance"
                                        ]
                                    },
                                    "serial": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Devices"
                ],
                "summary": "Delete a device",
                "operationId": "535264e834b177341eaa5eb39b2f3874",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
...
}